move hare/eligibility to hare3/eligibility
i initially thought that it will be replcaced with graded oracle fast
but it all takes time, and now it is just seems like a mess,
so i moved it to hare3 directory and adjusted imports
