+++ hare3/eligibility/interface.go
