+++ hare3/eligibility/mocks.go
