+++ hare3/eligibility/oracle.go
