+++ hare3/eligibility/oracle_test.go
